+++ tests/Functional/Command/CommandTestCase.php
@@ -23,14 +23,15 @@ abstract class CommandTestCase extends WebTestCase
     /**
      * Runs a command and returns it output.
      *
+     * The verbosity is always verbose because Symfony looks at the definition in the environment variables.
+     *
      * @param Client $client
      * @param string $command
-     * @param int    $verbosity Verbosity level to use
-     * @param int    $exitCode  Expected command exit code
+     * @param int    $exitCode Expected command exit code
      *
      * @return string
      */
-    protected function runCommand(Client $client, $command, $verbosity = OutputInterface::VERBOSITY_NORMAL, $exitCode = 0)
+    protected function runCommand(Client $client, $command, $exitCode = 0)
     {
         $application = new Application($client->getKernel());
         $application->setAutoExit(false);
@@ -38,24 +39,19 @@ abstract class CommandTestCase extends WebTestCase
         $fp = tmpfile();
         $input = new StringInput($command);
         $output = new StreamOutput($fp);
-        $output->setVerbosity($verbosity);
+        $output->setVerbosity(OutputInterface::VERBOSITY_DEBUG); // to work the same with Symfony < 4.0
 
         $realCode = $application->run($input, $output);
 
         fseek($fp, 0);
         $output = '';
         while (!feof($fp)) {
-            $output = fread($fp, 4096);
+            $output .= fread($fp, 4096);
         }
         fclose($fp);
 
         $this->assertEquals($exitCode, $realCode, $output);
 
         return $output;
-    }
-
-    protected function tearDown()
-    {
-        static::createClient()->getContainer()->unmock('fos_http_cache.cache_manager');
     }
 }