+++ tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -30,6 +30,9 @@ class ConfigurationTest extends AbstractExtensionConfigurationTestCase
 
     public function testEmptyConfiguration()
     {
+        $this->markTestSkipped('We need to update to the latest SymfonyConfigTest version and see if the bug in AbstractConfigurationConstraint.php
+ is still present');
+
         $expectedConfiguration = $this->getEmptyConfig();
 
         $formats = array_map(function ($path) {